requirements.txt
apache-airflow==2.10.2
pandas==2.2.3
apache-airflow-providers-docker==3.14.0